refactor(eudi): extract CGO-free holder-key seam into eudi/holderkeys
HolderSigner and the KB-JWT signer-keybinder bridge lived in eudi/wallet, but
that package imports sqlcipherstorage (cgo) via wallet.New — so any consumer of
the HolderSigner interface (the WSCA adapter, a server-side Postgres holder)
transitively pulled in sqlcipher and could not build with CGO_ENABLED=0.

Move the pure seam (HolderSigner, SoftwareHolderSigner, derToRawES256, the
NewSignerKeyBinder sdjwtvc.KeyBinder bridge) into a new CGO-free eudi/holderkeys
package, as #614's design doc intended. eudi/wallet keeps type aliases so its
public API is unchanged, and openid4vci.HolderKeyBinder was already CGO-free.
Consumers now import eudi/holderkeys and build CGO_ENABLED=0.
